+++ src/Search/SearchSubscriber.php
@@ -1,14 +1,24 @@
 <?php
 namespace Tui\PageBundle\Search;
 
-use Doctrine\Common\EventSubscriber;
-use Doctrine\ORM\Event\LifecycleEventArgs;
+use Doctrine\Bundle\DoctrineBundle\Attribute\AsDoctrineListener;
+use Doctrine\ORM\Event\PostPersistEventArgs;
+use Doctrine\ORM\Event\PreRemoveEventArgs;
 use Doctrine\ORM\Event\PreUpdateEventArgs;
+use Doctrine\ORM\Events;
 use Psr\Log\LoggerInterface;
 use Tui\PageBundle\Entity\IsIndexableInterface;
 use Tui\PageBundle\Entity\PageInterface;
 
-class SearchSubscriber implements EventSubscriber
+/**
+ * preUpdate -- get diff of availableLanguages, ensure indexes exist, add/remove translated document from each
+ * postPersist -- add new document to index
+ * preRemove -- delete from indexes
+ */
+#[AsDoctrineListener(event: Events::postPersist)]
+#[AsDoctrineListener(event: Events::preUpdate)]
+#[AsDoctrineListener(event: Events::preRemove)]
+class SearchSubscriber
 {
     private readonly bool $enabled;
     private ?array $collections = null;
@@ -22,19 +32,6 @@ class SearchSubscriber implements EventSubscriber
             return;
         }
         $this->enabled = $searchEnabled;
-    }
-
-    // preUpdate - get diff of availableLanguages, ensure indexes exist, add/remove translated document from each
-    // postPersist -- add new document to index
-    // preRemove -- delete from indexes
-
-    public function getSubscribedEvents(): array
-    {
-        return [
-            'postPersist',
-            'preUpdate',
-            'preRemove',
-        ];
     }
 
     public function preUpdate(PreUpdateEventArgs $args): void
@@ -61,7 +58,7 @@ class SearchSubscriber implements EventSubscriber
         }
     }
 
-    public function postPersist(LifecycleEventArgs $args): void
+    public function postPersist(PostPersistEventArgs $args): void
     {
         if (!$this->enabled) {
             return;
@@ -88,7 +85,7 @@ class SearchSubscriber implements EventSubscriber
         }
     }
 
-    public function preRemove(LifecycleEventArgs $args): void
+    public function preRemove(PreRemoveEventArgs $args): void
     {
         if (!$this->enabled) {
             return;